requirements.txt
streamlit>=1.30.0
Pillow>=10.0.0
pytesseract>=0.3.10
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
python-pptx>=0.6.21
reportlab>=4.0.0
numpy>=1.24.0
# hdbscan>=0.8.33  # Optional: Advanced clustering (falls back to K-Means if not available)
